requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-openai
langgraph
requests

